+++ src/App.tsx
@@ -11,7 +11,7 @@ import { PatientConditions } from "./components/content/conditions/patient-condi
 import { PatientMedicationsTabbed } from "./components/content/medications/patient-medications-tabbed";
 import { PatientAllergies } from "@/components/content/allergies/patient-allergies";
 import { PatientImmunizations } from "./components/content/immunizations/patient-immunizations";
-import { PatientCareTeam } from "./components/care-team/patient-careteam";
+import { PatientCareTeam } from "./components/content/care-team/patient-careteam";
 
 const {
   VITE_SYSTEM_URL,
