+++ src/index.ts
@@ -5,6 +5,7 @@ export * from "@/components/content/conditions";
 export * from "@/components/content/conditions/patient-conditions";
 export * from "@/components/content/conditions-table-base";
 export * from "@/components/content/immunizations/patient-immunizations";
+export * from "@/components/content/care-team/patient-careteam";
 export * from "@/components/content/medication-drawer";
 export * from "@/components/content/medications-table-base";
 export * from "@/components/content/medications/add-new-med-drawer";
